Make engine-tree independent of beacon-consensus crate
