+++ crates/consensus/beacon/src/engine/event.rs
@@ -1,6 +1,6 @@
-use crate::engine::forkchoice::ForkchoiceStatus;
 use alloy_primitives::B256;
 use alloy_rpc_types_engine::ForkchoiceState;
+use reth_engine_primitives::ForkchoiceStatus;
 use reth_primitives::{SealedBlock, SealedHeader};
 use std::{
     fmt::{Display, Formatter, Result},
