+++ crates/engine/primitives/Cargo.toml
@@ -16,9 +16,16 @@ reth-execution-types.workspace = true
 reth-payload-primitives.workspace = true
 reth-primitives.workspace = true
 reth-trie.workspace = true
+reth-errors.workspace = true
 
 # alloy
 alloy-primitives.workspace = true
+alloy-rpc-types-engine.workspace = true
+
+# async
+tokio = { workspace = true, features = ["sync"] }
+futures.workspace = true
 
 # misc
 serde.workspace = true
+thiserror.workspace = true
