+++ crates/engine/util/src/skip_new_payload.rs
@@ -2,8 +2,7 @@
 
 use alloy_rpc_types_engine::{PayloadStatus, PayloadStatusEnum};
 use futures::{Stream, StreamExt};
-use reth_beacon_consensus::BeaconEngineMessage;
-use reth_engine_primitives::EngineTypes;
+use reth_engine_primitives::{BeaconEngineMessage, EngineTypes};
 use std::{
     pin::Pin,
     task::{ready, Context, Poll},
